# api/openapi.yaml
openapi: 3.0.3
info:
  title: Trelay API
  description: |
    Trelay is a privacy-respecting, developer-first URL shortener.
    
    ## Authentication
    All endpoints (except health checks and redirects) require authentication via one of:
    - `X-API-Key` header with your API key
    - `Authorization: Bearer <jwt_token>` header with a JWT token
  version: 2.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/import:
    post:
      tags: [Import/Export]
      summary: Import links from CSV file
      operationId: importCSV
      security:
        - apiKey: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                format:
                  type: string
                  enum: [generic, yourls, shlink, bitly]
                skip_duplicates:
                  type: string
                  enum: ['true', 'false']
      responses:
        '200':
          description: Import result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/ImportResult'
  /api/v1/import/json:
    post:
      tags: [Import/Export]
      summary: Import links from JSON
      operationId: importJSON
      security:
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [links]
              properties:
                links:
                  type: array
                  items:
                    $ref: '#/components/schemas/ImportLink'
                skip_duplicates:
                  type: boolean
      responses:
        '200':
          description: Import result
  /api/v1/export:
    get:
      tags: [Import/Export]
      summary: Export all links
      operationId: exportLinks
      security:
        - apiKey: []
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [csv, json]
        - name: folder_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Export file
          content:
            text/csv:
              schema:
                type: string
            application/json:
              schema:
                type: object
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ImportLink:
      type: object
      required: [url]
      properties:
        url:
          type: string
        slug:
          type: string
        tags:
          type: array
          items:
            type: string
        password:
          type: string
        ttl_hours:
          type: integer
        folder_id:
          type: integer
    ImportResult:
      type: object
      properties:
        total:
          type: integer
        imported:
          type: integer
        skipped:
          type: integer
        failed:
          type: integer
        errors:
          type: array
          items:
            type: object
            properties:
              row:
                type: integer
              url:
                type: string
              slug:
                type: string
              message:
                type: string